+++ dgtpicom_dgt3000.h
@@ -75,8 +75,8 @@
 volatile unsigned *gpio, *gpioset, *gpioclr, *gpioin;
 volatile unsigned *i2cSlave, *i2cSlaveRSR, *i2cSlaveSLV, *i2cSlaveCR, *i2cSlaveFR;
 volatile unsigned *i2cMaster, *i2cMasterS, *i2cMasterDLEN, *i2cMasterA, *i2cMasterFIFO, *i2cMasterDiv, *i2cMasterDel;
-int32_t *timerh;
-int32_t *timerl;
+u_int32_t *timerh;
+u_int32_t *timerl;
 
 // variables for debug stats
 #ifdef debug
@@ -188,7 +188,7 @@ const char crc_table[256] = {
 	2 = Pi 2 */
 int checkPiModel();
 
-long long int * timer();
+u_int64_t * timer();
 
 int checkCoreFreq();
 
@@ -291,5 +291,5 @@ void *dgt3000Receive(void *);
 	returns:
 	-3 = no Ack
 	0 = Ack */
-int dgt3000GetAck(char adr, char cmd, long long int timeOut);
+int dgt3000GetAck(char adr, char cmd, u_int64_t timeOut);
